requirements.txt
nicegui>=2.0.0
qdrant-client>=1.14.0
SQLAlchemy>=2.0
psycopg2-binary>=2.9
python-dotenv>=1.0
requests>=2.31
python-docx>=1.1
beautifulsoup4>=4.12
lxml>=5.0
